+++ rn_messaging_bot/models/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from . import rn_messaging_template
+from . import rn_messaging_campaign
 from . import rn_messaging_connector
 from . import rn_messaging_conversation
 from . import rn_messaging_message
@@ -9,6 +11,8 @@ from . import rn_messaging_bot_session
 from . import res_partner
 
 from ..services import bot_engine_service
+from ..services import campaign_service
 from ..services import connector_service
 from ..services import queue_service
+from ..services import template_service
 from ..services import webhook_service